+++ src/dnsbl_check/provider.py
@@ -1,108 +1,78 @@
-# pylint: disable=W0401,W0614
-from config import *
+from pycares import ares_query_a_result
+
+from config import DNSBL_CATEGORY_UNKNOWN
 
 
-class Provider:
-    def __init__(self, host: str):
-        self.host = host
+class BaseProvider:
+    IP4 = False
+    IP6 = False
+    DOMAIN = False
+    HOST = None
 
-    def process_response(self, response):
-        """
-        Usually DNSBL lists returns ip-codes like this: 127.0.0.2
-        Some of the lists provides some specification about
-        this codes and their meaning. This function will be helpful to build mapping
-        between response codes and DNSBL_CATEGORIES.  It is used in construction
-        of DNSBLResult. You should redefine this function
-        in your custom providers according to their specification.
+    RESPONSE_CATEGORIES = {}
 
-        Parmeters:
-            result - list of c-ares dns responses
+    def __init__(self, host: str = None):
+        if self.HOST is None and host is None:
+            raise ValueError('The provider hostname needs to be provided')
 
-        Returns:
-            set of categories (DNSBL_CATEGORIES subset)
+        if self.HOST is not None:
+            self.host = self.HOST
 
-        """
-        result = set()
-        if response:
-            result.add(DNSBL_CATEGORY_UNKNOWN)
+        else:
+            self.host = host
 
-        return result
+    def response_categories(self, response: (list[ares_query_a_result], ares_query_a_result)) -> set[str]:
+        categories = set()
+
+        if not response:
+            return categories
+
+        if len(self.RESPONSE_CATEGORIES) == 0:
+            categories.add(DNSBL_CATEGORY_UNKNOWN)
+
+        else:
+            if not isinstance(response, list):
+                response = [response]
+
+            for res in response:
+                cat = self.RESPONSE_CATEGORIES.get(res.host, DNSBL_CATEGORY_UNKNOWN)
+                if isinstance(cat, set):
+                    categories.update(cat)
+
+                else:
+                    categories.add(cat)
+
+        return categories
 
     def __repr__(self):
         return f"<Provider: {self.host}>"
 
 
-class ZenSpamhaus(Provider):
-    """ Combined spamhaus list:
-        https://www.spamhaus.org/faq/section/DNSBL%20Usage#200
-    """
-
-    def __init__(self):
-        Provider.__init__(self, host='zen.spamhaus.org')
-
-    def process_response(self, response):
-        categories = set()
-        for result in response:
-            if result.host in ['127.0.0.2', '127.0.0.3', '127.0.0.9']:
-                categories.add(DNSBL_CATEGORY_SPAM)
-
-            elif result.host in ['127.0.0.4', '127.0.0.5', '127.0.0.6', '127.0.0.7']:
-                categories.add(DNSBL_CATEGORY_EXPLOITS)
-
-            elif result.host in ['127.255.255.252', '127.255.255.254', '127.255.255.255']:
-                categories.add(DNSBL_CATEGORY_ERROR)
-
-            else:
-                categories.add(DNSBL_CATEGORY_UNKNOWN)
-
-        return categories
+class Provider(BaseProvider):
+    IP4 = True
+    IP6 = True
+    DOMAIN = True
 
 
-class DblSpamhaus(Provider):
-    """ Spamhaus domain blacklist
-        https://www.spamhaus.org/faq/section/Spamhaus%20DBL#291
-    """
-    CATEGORY_MAPPING = {
-        '127.0.1.2': {DNSBL_CATEGORY_SPAM},
-        '127.0.1.4': {DNSBL_CATEGORY_PHISH},
-        '127.0.1.5': {DNSBL_CATEGORY_MALWARE},
-        '127.0.1.6': {DNSBL_CATEGORY_CNC},
-        '127.0.1.102': {DNSBL_CATEGORY_ABUSED, DNSBL_CATEGORY_LEGIT, DNSBL_CATEGORY_SPAM},
-        '127.0.1.103': {DNSBL_CATEGORY_ABUSED, DNSBL_CATEGORY_SPAM},
-        '127.0.1.104': {DNSBL_CATEGORY_ABUSED, DNSBL_CATEGORY_LEGIT, DNSBL_CATEGORY_PHISH},
-        '127.0.1.105': {DNSBL_CATEGORY_ABUSED, DNSBL_CATEGORY_LEGIT, DNSBL_CATEGORY_MALWARE},
-        '127.0.1.106': {DNSBL_CATEGORY_ABUSED,  DNSBL_CATEGORY_LEGIT, DNSBL_CATEGORY_CNC},
-        '127.255.255.252': {DNSBL_CATEGORY_ERROR},
-        '127.255.255.254': {DNSBL_CATEGORY_ERROR},
-        '127.255.255.255': {DNSBL_CATEGORY_ERROR},
-    }
-
-    def __init__(self):
-        Provider.__init__(self, host='dbl.spamhaus.org')
-
-    def process_response(self, response):
-        categories = set()
-        for result in response:
-            result_categories = self.CATEGORY_MAPPING.get(result.host, {DNSBL_CATEGORY_UNKNOWN})
-            categories.update(result_categories)
-
-        return categories
+class ProviderIP(BaseProvider):
+    IP4 = True
+    IP6 = True
+    DOMAIN = False
 
 
-CUSTOM_PROVIDERS = {
-    'zen.spamhaus.org': ZenSpamhaus,
-    'dbl.spamhaus.org': DblSpamhaus,
-}
+class ProviderIP4(BaseProvider):
+    IP4 = True
+    IP6 = False
+    DOMAIN = False
 
-BASE_PROVIDERS_IP = [
-     Provider(host) for host in RAW_PROVIDERS_IP if host not in CUSTOM_PROVIDERS
-]
-BASE_PROVIDERS_IP.extend([
-     CUSTOM_PROVIDERS[host]() for host in RAW_PROVIDERS_IP if host in CUSTOM_PROVIDERS
-])
-BASE_PROVIDERS_DOMAIN = [
-     Provider(host) for host in RAW_PROVIDERS_DOMAIN if host not in CUSTOM_PROVIDERS
-]
-BASE_PROVIDERS_DOMAIN.extend([
-     CUSTOM_PROVIDERS[host]() for host in RAW_PROVIDERS_DOMAIN if host in CUSTOM_PROVIDERS
-])
+
+class ProviderIP6(BaseProvider):
+    IP4 = False
+    IP6 = True
+    DOMAIN = False
+
+
+class ProviderDomain(BaseProvider):
+    IP4 = False
+    IP6 = False
+    DOMAIN = True